move makeUserObject and add new options
