+++ src/applications/personalization/dashboard/tests/e2e/dashboard-e2e-helpers.js
@@ -1,61 +1,11 @@
+// makeUserObject used to be defined in this file. To avoid updating all places
+// that were already using it, we're just re-exporting it
+export { makeUserObject } from '~/applications/personalization/common/helpers';
+
 export function mockLocalStorage() {
   // make sure no first-time UX modals are in the way
   window.localStorage.setItem(
     'DISMISSED_ANNOUNCEMENTS',
     JSON.stringify(['single-sign-on-intro', 'find-benefits-intro']),
   );
-}
-export function makeUserObject(options = {}) {
-  const services = [];
-  if (options.rx) {
-    services.push('rx');
-  }
-  if (options.messaging) {
-    services.push('messaging');
-  }
-  return {
-    data: {
-      id: '',
-      type: 'users_scaffolds',
-      attributes: {
-        services,
-        account: { accountUuid: 'c049d895-ecdf-40a4-ac0f-7947a06ea0c2' },
-        profile: {
-          email: 'vets.gov.user+36@gmail.com',
-          firstName: 'WESLEY',
-          middleName: 'WATSON',
-          lastName: 'FORD',
-          birthDate: '1986-05-06',
-          gender: 'M',
-          zip: '21122-6706',
-          lastSignedIn: '2020-07-21T00:04:51.589Z',
-          loa: { current: 3, highest: 3 },
-          multifactor: true,
-          verified: true,
-          signIn: { serviceName: 'idme', accountType: 'N/A', ssoe: true },
-          authnContext: 'http://idmanagement.gov/ns/assurance/loa/3',
-        },
-        vaProfile: {
-          status: 'OK',
-          birthDate: '19860506',
-          familyName: 'Ford',
-          gender: 'M',
-          givenNames: ['Wesley', 'Watson'],
-          isCernerPatient: options.isCerner,
-          facilities: options.facilities || [],
-          vaPatient: options.isPatient,
-          mhvAccountState: 'NONE',
-        },
-        veteranStatus: {
-          status: 'OK',
-          isVeteran: true,
-          servedInMilitary: true,
-        },
-        inProgressForms: options.inProgressForms || [],
-        prefillsAvailable: [],
-        vet360ContactInformation: {},
-      },
-    },
-    meta: { errors: null },
-  };
 }
